+++ src/MicroElements.Metadata/Utils/CodeCompiler.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Concurrent;
 using System.Linq.Expressions;
 using System.Reflection;
-using MicroElements.Functional;
 
 namespace MicroElements
 {